requirement.txt
pytest
PyPDF2
python-docx
ollama
openai
